- BUGFIX: crash on second backup run - ADD: build directory - ADD: Microsoft Visual C++ 2008 Express Edition Project (doesn't work correct yet)

+++ trunk/src/BFCronCtrl.cpp
@@ -99,7 +99,7 @@ void BFCronCtrl::GetData()
     pComboType_->Select(iType);
     FillCombos();
 
-    if ( (lVal-1) > pComboIntervall_->GetCount() )
+    if ( (lVal-1) > (long)pComboIntervall_->GetCount() )
         return;
 
     pComboIntervall_->SetSelection(lVal-1);
